+++ Sources/CapsuleSwift/CapsuleSessionState.swift
@@ -1,0 +1,15 @@
+//
+//  File.swift
+//  
+//
+//  Created by Joe Blau on 6/25/24.
+//
+
+import Foundation
+
+public enum CapsuleSessionState: Int {
+    case unknown
+    case inactive
+    case active
+    case activeLoggedIn
+}
